Split out the BabySqueel::ActiveRecord namespace into separate files
